Add files via upload
In this project the Key Challenges were :
1) Queue Over Crowding :- Unorganized queue formation leading to inefficiency.
2) Poor Communication :- Students unaware of queue status and order readiness.
3) Time Estimation Issues :- No accurate waiting time prediction affects students' schedules.

So basically what we have observed that the Lunch break timings of all the students is same and there are a specific servers(2-3) people serving thali at a time creating a long queue and ding changes in time isn't just the Solution. So by getting inspired by the RFID system at Decathlon Showrooms we thought that already a student has a normal ID card, why not RFID?? if it becomes RFID it becomes easy to track and execute.

Solution : So firstly what do we need? a) Proper Interaction system -- At the entry of the canteen and At the buffet counter while collecting the thalis. (ESP8266 + RFID reader)
                Secondly we need a proper Database so we are using MongoDB here. 
                Thirdly we need is a proper Backend to properly function between the hardware and the user. (Python -- Flask)
                 Fourthly we need a proper communication System So that Overcrowding can be dealt. ( Twilio API's)